I want `Integer#divisors` that return the divisors of `self`.
```ruby
6.divisors # => [1, 2, 3, 6]
7.divisors # => [1, 7]
8.divisors # => [1, 2, 4, 8]
```
Divisors are basic and important mathematical matters.
Enumeration of divisors can be useful or necessary when solving integer problems.

The main idea and `divisors` code was provided by tompng-san.
I improved the average speed and added document and tests.